requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-groq
arxiv
wikipedia
## tracking/evaluation of ur apps
langsmith
langchain_tavily

langchain-mcp-adapters
mcp
langgraph-cli[inmem]